test compat

+++ emoji_script/tests/samples/14.py
@@ -6,9 +6,9 @@ def median(pool):
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
-        return copy[(size - 1) / 2]
+        return copy[int((size - 1) / 2)]
     else:
-        return (copy[size/2 - 1] + copy[size/2]) / 2
+        return (copy[int(size / 2 - 1)] + copy[int(size / 2)]) / 2
 if __name__ == '__main__':
     import doctest
     doctest.testmod()
